+++ packages/terra-search-field/CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Changed
+* Add ability to default value and onChange function handler
 * Add ability to disable and stretch to fill its container
 * Add ability to disable the auto searching and trigger searching via ENTER key